# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Financial Constraint Analysis Platform API
  description: Enterprise-grade financial analysis platform with AI-powered insights, constraint management, and comprehensive compliance features.
  version: 1.0.0
  contact:
    name: Financial Analyzer Team
    email: support@financial-analyzer.com
  license:
    name: Proprietary
    url: https://financial-analyzer.com/license
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.financial-analyzer.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /privacy/export-request:
    post:
      summary: Request data export (GDPR Article 15)
      tags: [Privacy & GDPR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestType]
              properties:
                requestType:
                  type: string
                  enum: [PERSONAL_DATA, ACTIVITY_LOG, DOCUMENTS, ANALYSIS_RESULTS, FULL_EXPORT]
                options:
                  type: object
                  properties:
                    format:
                      type: string
                      enum: [JSON, CSV, XML]
                      default: JSON
                    anonymize:
                      type: boolean
                      default: false
      responses:
        '202':
          description: Export request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requestId:
                    type: string
                  message:
                    type: string
                  estimatedCompletion:
                    type: string
  /privacy/deletion-request:
    post:
      summary: Request data deletion (GDPR Article 17)
      tags: [Privacy & GDPR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deletionType]
              properties:
                deletionType:
                  type: string
                  enum: [ACCOUNT_DELETION, DATA_PURGE, DOCUMENT_DELETION, ANALYSIS_DELETION, PARTIAL_DELETION]
                options:
                  type: object
                  properties:
                    retainAuditLogs:
                      type: boolean
                      default: true
                    reason:
                      type: string
      responses:
        '202':
          description: Deletion request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requestId:
                    type: string
                  message:
                    type: string
                  warningPeriod:
                    type: string
  /privacy/consent:
    post:
      summary: Record user consent
      tags: [Privacy & GDPR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consentType, granted]
              properties:
                consentType:
                  type: string
                  enum: [privacy_policy, terms_of_service, marketing, data_retention]
                version:
                  type: string
                  default: "1.0"
                granted:
                  type: boolean
      responses:
        '200':
          description: Consent recorded successfully

  # Security
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...